requirements.txt
alembic==1.4.2
autopep8==1.5.4
boto3==1.14.38
botocore==1.17.38
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
